move handlers to its own package
