+++ internal/api/handler/auth.go
@@ -1,4 +1,4 @@
-package api
+package handler
 
 import (
 	"context"
