+++ internal/api/handler/auth_helpers.go
@@ -1,4 +1,4 @@
-package api
+package handler
 
 import (
 	"context"
